split phase from mod
